requirements_apple_vision.txt
astor==0.8.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
decorator==5.1.0
dill==0.3.4
filesplit==3.0.2
fonttools==4.32.0
future==0.18.2
gast==0.3.3
idna==3.3
imageio==2.13.3
imgaug==0.4.0
kiwisolver==1.3.2
lmdb==1.2.1
matplotlib==3.5.1
multiprocess==0.70.12.2
networkx==2.6.3
Nuitka==0.7.7
numpy==1.19.3
opencc-python-reimplemented==0.1.6
opencv-python==4.5.4.60
packaging==21.3
paddlepaddle==2.2.2
pefile==2021.9.3
Pillow==8.4.0
protobuf==3.19.1
pyclipper==1.3.0.post2
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-Vision==8.5
pyparsing==3.0.6
PySimpleGUI==4.55.1
pysrt==1.1.2
python-dateutil==2.8.2
python-Levenshtein-wheels==0.13.2
PyWavelets==1.2.0
QPT==1.0b3.dev7
requests==2.26.0
scikit-image==0.19.1
scipy==1.7.3
Shapely==1.8.0
six==1.16.0
tifffile==2021.11.2
tk==0.1.0
tqdm==4.63.1
ttkbootstrap==0.5.1
urllib3==1.26.7
wget==3.2
whych==0.2.1
wordsegment==1.3.1
wurlitzer==3.0.2
zstandard==0.17.0
